requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Database and ORM (if needed beyond Supabase)
sqlalchemy==2.0.23
alembic==1.13.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Supabase client
supabase==2.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3

# Redis and task queue
redis==5.0.1
celery==5.3.4

# HTTP client
httpx==0.24.0

# AI/ML libraries
openai==1.3.7

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
PyJWT==2.9.0

# Stripe
stripe==7.0.0